% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cc}
\alias{cc}
\title{Cockcroft–Gault formula for creatinin clearance calculation}
\usage{
cc(age, weight, sex, scre)
}
\arguments{
\item{age}{is the anagraphical age of the patients expressed in years}

\item{weight}{is the patients' mass expressed in kilograms}

\item{sex}{is the gender of the patients expresed as 0 for Male and 1 for Female}

\item{scre}{is the serum creatinin expressed in mg/dL units.}
}
\value{
cc.out is vector of the creatinine clearance.
}
\description{
Cockcroft–Gault formula for creatinin clearance calculation
}

