% Generated by roxygen2 (4.0.2): do not edit by hand
\name{EuroScore_I_Log}
\alias{EuroScore_I_Log}
\title{EuroScore is a mortality estimation model used in cardiovascular medicine, particularly in cardiac surgery.}
\usage{
EuroScore_I_Log(x, ...)
}
\arguments{
\item{x}{is datframe with all variables included in orginal score from 1999}
}
\value{
returns additive score added othe data frame wiht new variabel of Logistic EuroScore.
}
\description{
EuroScore is a mortality estimation model used in cardiovascular medicine, particularly in cardiac surgery.
}

