% Generated by roxygen2 (4.0.2): do not edit by hand
\name{EuroScoreLog.II}
\alias{EuroScoreLog.II}
\title{EuroScore II is a mortality estimation model used in cardiovascular medicine, particularly in cardiac surgery.}
\usage{
EuroScoreLog.II(x, ...)
}
\arguments{
\item{x}{is datframe with all variables included in orginal score from 2008}
}
\value{
returns additive score added othe data frame wiht new variabel of Logistic EuroScore II.
}
\description{
EuroScore II is a mortality estimation model used in cardiovascular medicine, particularly in cardiac surgery.
}

