% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gfr_ckd}
\alias{gfr_ckd}
\title{Glomerular filatration rate by CKD-EPI}
\usage{
gfr_ckd(gender = "female|male", scr = NULL, scr.u = "mg/dl" | "µmol/l",
  age = NULL, race = "black|white")
}
\description{
Glomerular filatration rate by CKD-EPI
}

